Refactor caches
